make validation work
