+++ components/admin/Table/useTable.ts
@@ -8,7 +8,8 @@ import {
 } from '@tanstack/react-table'
 import type { player, tournament, match } from '@prisma/client'
 
-import { FORM_VALUES, DEFAULT_PAGINATION } from 'constants/values'
+import { DEFAULT_PAGINATION } from 'constants/values'
+import { FORM_VALUES } from 'constants/formValues'
 import type { PaginationProps } from '../Pagination'
 
 const columnHelper = createColumnHelper()
